refactor: extract microlearning list related actions to separate components
